Temporary coroutine patch (pending illumos 15206)
The coroutine ucontext backend uses a combination of sigsetjmp/siglongjmp
and set/getcontext(). It stores the current coroutine data in thread local
storage. When a context is restored into a different thread than the one in
which it last ran, the %fsbase is also restored, leading to a thread that
thinks it is a different one, and critically for this code, a thread that
accesses the TLS of the other.

This patch copies %fsbase from the existing thread context to the new one
before switching so that it resumes with the right identity. If relies on this
being an amd64 binary, and insider knowledge of the libc implementations.

It is not yet understood why this code works on other platforms without this
change, or if there is a better fix. It would be slightly cleaner to use
set/getcontext() throughout, but that adds additional system call overhead.

For more analysis, see https://github.com/omniosorg/illumos-omnios/issues/1260